refactor: remove all sqlfluff only test cases to test_sqlfluff.py
